+++ src/main/java/racingcar/controller/MovingStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.controller;
 
 public interface MovingStrategy {
