+++ src/main/java/racingcar/controller/RacingCarCommander.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.controller;
 
 import java.util.Random;
