+++ src/main/java/racingcar/domain/RacingCars.java
@@ -1,10 +1,9 @@
 package racingcar.domain;
 
 import java.util.Arrays;
-import java.util.Comparator;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
+import racingcar.controller.MovingStrategy;
 import racingcar.controller.RacingCarDto;
 
 public class RacingCars {
