+++ src/test/java/racingcar/domain/RacingCarsTest.java
@@ -9,6 +9,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EmptySource;
+import racingcar.controller.MovingStrategy;
+import racingcar.controller.RacingCarCommander;
 
 class RacingCarsTest {
