+++ src/test/java/racingcar/domain/RacingGameTest.java
@@ -4,15 +4,12 @@ import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import java.util.Scanner;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.controller.RacingCarController;
-import racingcar.view.InputView;
-import racingcar.view.OutputView;
+import racingcar.controller.RacingCarCommander;
 
 public class RacingGameTest {
