+++ src/test/java/racingcar/domain/RandomGeneratorTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
+import racingcar.controller.RacingCarCommander;
 
 public class RandomGeneratorTest {
